backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI and ML Libraries
google-generativeai==0.3.2
opencv-python==4.8.1.78
mediapipe==0.10.21
pyaudioanalysis==0.3.14
numpy==1.24.3
scipy==1.11.4

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Email
fastapi-mail==1.4.1

# WebSocket support
websockets==12.0

# Validation
pydantic==2.5.0
pydantic-settings==2.2.1
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
python-dotenv==1.0.0